fix: image imports on login and signup pages (#49)
* fix: image imports on login and signup pages

* refactor/separate-components-mutations (#53)

* feat: added vitest and updated pre-commit hooks and justfile

* feat: moved login page and mutation into separate files

* feat: moved signup into different files

* feat: login mutation uses cookies instead of local storage
