+++ frontend/package.json
@@ -44,7 +44,7 @@
     "prettier": "^3.5.3",
     "typescript": "^5.9.3",
     "vite": "^7.1.7",
-    "vitest": "^3.0.5",
+    "vitest": "^3.2.4",
     "web-vitals": "^5.1.0"
   }
 }
